Address panel review findings on the bundled ConPTY change
Three-model panel (Codex + GLM + Claude subagent) review of bef35e2b:

- Split the ConPTY helpers into shared/conptyWindows.ts. platform.ts
  reads process.platform at module scope, which the sandboxed renderer
  cannot evaluate (measured: preload.ts documents it, pathLinkProvider
  guards with typeof process) — importing it from useTerminal.ts risked
  a ReferenceError that would take the whole renderer bundle down.
  conptyWindows.ts is pure parameters-in/answer-out, browser-safe.
- Wrap the in-box fallback retry in its own try/catch so a failing
  retry still surfaces the actionable 'Failed to start shell' shape
  instead of a raw node-pty throw.
- Log the backend on every gated spawn success, not only on fallback —
  the changelog promises it and 'shipped, still broken' reports are
  undiagnosable without it.
- Also treat 'Cannot launch conpty' (conpty.cc:307, pseudoconsole
  creation failed e.g. antivirus blocking OpenConsole.exe) as a
  fallback trigger: without it a damaged install hard-fails the shell,
  strictly worse than before the fix.

Accepted as-is (owner decisions): the fallback path can leave xterm on
the modern-build heuristic while the PTY runs in-box (R1 — rare
packaging failure only, diagnosable via the new logs), and node-pty
1.1.0's DLL mode skips conout-worker cleanup on natural exit (R2 —
observed via the dogfood checklist, upstream issue if confirmed).
