+++ changelog.d/910.md
@@ -1,3 +1,3 @@
 ### Fixed
 
-- **Mouse clicks and wheel now reach vim and other TUIs on Windows 10.** The in-box ConPTY on Windows 10 (and Server 2022) never forwards the escape sequences that turn on mouse reporting, so `set mouse=a` in vim silently did nothing — the wheel scrolled wmux's own scrollback instead. Those builds now spawn their panes against the OpenConsole build bundled with wmux, which forwards mouse events correctly. Windows 11 is untouched. If the bundled backend cannot load (a damaged install), wmux logs which backend actually started and falls back to the previous behaviour, so the shell still opens.
+- **Mouse clicks and wheel now reach vim and other TUIs on Windows 10.** The in-box ConPTY on Windows 10 (and Server 2022) never forwards the escape sequences that turn on mouse reporting, so `set mouse=a` in vim silently did nothing — the wheel scrolled wmux's own scrollback instead. Those builds now spawn their panes against the OpenConsole build bundled with wmux, which forwards mouse events correctly. Windows 11 is untouched. Every spawn logs which ConPTY backend actually started, and if the bundled one cannot load (a damaged install) wmux falls back to the previous behaviour, so the shell still opens.
