+++ src/daemon/DaemonSessionManager.ts
@@ -12,7 +12,8 @@ import { buildSpawnInjection, classifyShell } from './shell-integration';
 import { expandTilde } from '../shared/expandTilde';
 import { buildExecArgs } from './execWrapper';
 import { buildSafeChildEnv } from '../shared/envFilter';
-import { isMac, parseWindowsBuildNumber, shouldUseBundledConpty, isConptyDllLoadError } from '../shared/platform';
+import { isMac, parseWindowsBuildNumber } from '../shared/platform';
+import { shouldUseBundledConpty, isConptyDllLoadError } from '../shared/conptyWindows';
 import { getWindowsDefaultShell, resolveBareShellName, resolveLaunchableWindowsExe } from '../shared/shellResolution';
 import { ENV_KEYS } from '../shared/constants';
 import { createDefaultConfig } from './config';
@@ -440,6 +441,12 @@ export class DaemonSessionManager extends EventEmitter {
         useConpty: true,
         ...(useConptyDll ? { useConptyDll: true } : {}),
       });
+      if (useConptyDll) {
+        // #910 dogfood: without this line there is no way to tell whether a
+        // "shipped, still broken" report means the gate never fired or the
+        // bundled backend itself failed on that machine.
+        console.log(`[DaemonSessionManager] session ${params.id}: ConPTY backend = bundled conpty.dll`);
+      }
     } catch (err) {
       const detail = err instanceof Error ? err.message : String(err);
       if (useConptyDll && isConptyDllLoadError(detail)) {
@@ -448,14 +455,19 @@ export class DaemonSessionManager extends EventEmitter {
         // once against the in-box backend (mouse stays broken, but the shell
         // starts — same behaviour as before this fix).
         console.error(`[DaemonSessionManager] bundled conpty.dll failed to load (${detail}); falling back to in-box ConPTY`);
-        ptyProcess = pty.spawn(cmd, spawnArgs, {
-          name: 'xterm-256color',
-          cols,
-          rows,
-          cwd,
-          env,
-          useConpty: true,
-        });
+        try {
+          ptyProcess = pty.spawn(cmd, spawnArgs, {
+            name: 'xterm-256color',
+            cols,
+            rows,
+            cwd,
+            env,
+            useConpty: true,
+          });
+        } catch (retryErr) {
+          const retryDetail = retryErr instanceof Error ? retryErr.message : String(retryErr);
+          throw new Error(`Failed to start shell "${cmd}" in "${cwd}" (in-box ConPTY retry): ${retryDetail}`);
+        }
       } else {
         throw new Error(`Failed to start shell "${cmd}" in "${cwd}": ${detail}`);
       }
